Fix: libcrmcommon: Correctly delete XML comments according to their positions
Previously, v2 patchset was not able to define which exact XML comments were
expected to be deleted in a CIB XML element. It was a problem if there
was more than one comment at the same level within a CIB XML element.

This commit resolves it by adding and handling a "position" field in the
diff operation "delete" for XML comments.