Fix: -Wformat-truncation=2: treat getpid() string formatting generously
Discovered thanks to -Wformat-truncation=2 of GCC7.

We now stick with a convention used elsewhere in the code: pid_t is
treated as if it was unsigned long as it is, overall, better to upcast
than truncate and there is no handy printf conversion specifier for
this very type.  This plays it relatively safe for the foreseeable
future.

At that opportunity, also remove the respective superfluous initial
string initialization/termination + deal a bit better with OOM situation
in one instance.