Fix: tools: crm_resource -C should clear last-failure
Previously, this only cleared fail-count attributes. Now, it also clears
last-failure, which is better (and probably more expected) behavior.
This is accomplished by using attrd's new clear-failure op, which will also
make future changes to failure handling easier.

This has a small backward compatibility issue: if crm_resource -C is run
on a newer node in a mixed-version cluster that uses legacy attrd,
it will be unable to clean failures on an older node. Since we are bumping
CRM_FEATURE_SET with this release, this issue will be limited to the
rolling upgrade window. Note that this does not affect atomic attrd or
cleanups initiated by the DC's crmd.