Fix: libfencing,fencing: properly remap "action" in configuration
"action" should never be specified in fence device configuration.

Previously, this was handled by re-inserting the proper action when creating a
fence command, if the proper action were on a "safe" list (list, monitor,
status, or meta-data).

However, that was insufficient, partly because "on" should have been on the
list, but also because action="reboot" would interfere with reboots that were
remapped to off or off+on.

Now, stonithd intelligently maps any action parameter to pcmk_off_action and/or
pcmk_reboot_action as appropriate when the device configuration is registered,
and libfencing ignores any action parameter that makes it that far.