Feature: tools: make crm_failcount compatible with per-operation failcounts
This adds -n/--operation and -I/--interval options to specify a
particular operation to query or clear the failcount for.
If not specified, the command works on all operations and intervals;
for querying, this is accomplished by querying the CIB directly
and summing the results.

crm_failcount is perhaps complicated enough now to merit converting to C.
However, it needs to know the local node name, which currently means it would
need to be linked against libcrmcluster, and that is undesirable.