Fix: crmd,libcrmcommon: update throttle when CPUs are hot-plugged
Previously, the number of CPU cores was determined the first time it was
needed, and remembered permanently after that. That becomes inaccurate when
cores are hot-plugged in and out of a virtual machine. Now, the number of cores
is parsed each time it is needed (using the new libcrmcommon function).