XML: upgrade-1.3.xsl: simplify branching
Previously, the branching -- so as to map ACL target definition as
present with read/write/deny element in 1.x schemas into the definition
as present with acl_permission element in 2.x schemas -- was unjustly
overcomplicated.  Thanks to partial mutual exclusion of attributes
encoding such definition (only "attribute" is an add-in for all of them)
in the preimage, we can utilize more straightforward branching within
single <xsl:choose>, which makes the idea more easier to grok.