Fix: libpe_status: correctly flag remote connection resources
Previously, unpack_resources() would set is_remote_node (when appropriate)
after calling common_unpack(). However, this meant that other paths to
common_unpack() (such as for bundles) did not get this set correctly.

Now, is_remote_node is set in common_unpack() so all paths are correct.
It's also the logical place to do it, since it really is part of unpacking,
and saves an extra call to xml_contains_remote_node().