Low: attrd: broadcast local overrides of sync'ed attributes
In the case of an attribute that was set on a stopped node, when the node
rejoins, the peer writer will sync its attributes including that value. The
node will reject it, overriding it with the local value (empty string).

Previously, it would send that override only to the peer writer, meaning that
any other nodes would still have the previous value in memory (and would
respond to queries with it). Now, the node broadcasts any local overrides,
so that all nodes get the correct value.