Low: libcib: always use current values when unpacking config
Previously, cib_read_config() called unpack_instance_attributes() with
overwrite=FALSE. This meant that changes to an option would not take effect
unless the option was not set before.

The only significant use of cib_read_config() was in cib_acl_enabled(), which
used a new, empty hash table for every call, so the issue didn't matter.

The cib daemon also used cib_read_config() to maintain a global config_hash,
which was affected by the issue, but didn't matter because it was never used.
This change will allow config_hash to be used.