XML: upgrade-1.3.xsl: be explicit about non 1:1 wrt. ACL @attributes
This makes it explicit that/when the conversion is not loseless
(unfortunately, so far, there was an implicit but false understanding
that going from 1.x version of schema to 2.x line is perfectly
smooth, which is not the case with ACLs when @attribute was used
with read/write/deny elements -- add also a note about possible
future refinement).

We can use <xsl:message> to emit diagnotics about the affected corner
cases, which is now directed into standard logging subsystem as of
the previous commit (4e32bd7).