Fix: pengine: ignore transitive anti-colocations of anti-colocations
Fix for CLBZ#5320

Consider:

   dummy1 avoids node1
   dummy1 anti-colocates with dummy2
   dummy1 anti-colocates with dummy3
   dummy2 anti-colocates with dummy3

When considering node1 while placing dummy3, the PE correctly ignores dummy1's
score for node1, because both the anti-colocation and the node preference are
negative (per comment "negative preference for a node with a negative score
should not become a positive preference").

Then the PE takes into account dummy2's preferences. Its direct score for node1
is ignored for the same reason. But then the PE takes into account the
transitive constraints from dummy2 (i.e. dummy1 anti-colocates with dummy2).

Before this change, the PE would follow "the enemy of my enemy is my friend",
and dummy3 would absorb dummy1's avoidance of node1, because dummy1 and dummy3
both conflict with dummy2.

This makes some sense, with the idea that we'll have the best chance of placing
all resources if the transitive avoidance is taken into account. However, it is
undesirable in some situations -- for example, even if dummy3 has a preference
for node1, dummy1's -INFINITY preference for node1 would take precedence, when
it clearly shouldn't.

Now, transitive preferences are ignored if both colocations involved are
anti-colocations.