Fix: attrd,crmd: erase attributes at attrd start-up, not first join
Previously, the crmd cleared the local node's transient attributes from the CIB
the first time it joined the cluster, to handle the case where a node restarts
so quickly that the cluster layer doesn't notice.

However, that left a race where any attributes set between attrd's start-up and
crmd's first join would be lost.

Now, attrd handles the clearing when it starts. This makes more sense, since
attrd's attribute memory is empty at this point, and it is making sure the CIB
is in sync.