Low: crmd,pengine,libpe_status,libpe_rules: remove support for versioned attributes
This is still not ready for production. The main issue is that meta-data
cannot yet be reliably collected asynchronously from Pacemaker Remote nodes.
There is a plan to do this by scheduling a new CRM operation to gather
all required meta-data at the start of each transition, but it is not ready.