Low: tools: partially revert ccbdb2a2 to avoid attrd_updater needing libcrmcluster
Both approaches have drawbacks when used on remote nodes.

If we keep the get_local_node_name() calls, attrd_udpater needs libcrmcluster,
which shouldn't be required on remote nodes. Also, on a corosync cluster for
example, the call on a remote node will spend 10+ seconds trying to connect to
corosync before finally using uname, which will be incorrect for remote nodes
whose name known to the cluster is not the same as uname.

If we avoid get_local_node_name() (the approach here), there is not a clear way
to distinguish a query with a NULL node name meaning "let the remote proxy
callback add the name" from "query all nodes". However, this likely exists
already, and should be addressed separately.