Fix: libcrmservice: parse LSB long description correctly
Previously, we would use only the long description continuation lines, ignoring
the part of the long description on the description line itself. Also, in what
appears to be a misguided attempt to push a read line back into the file stream,
we would try writing to a read-only file pointer.