Fix: PE: Do not probe connection resources until the container is active
Otherwise we cannot rely on the REMOTE_CONTAINER_HACK logic for
replacing 'addr' with the node where the container is running.

This leads to monitor operations that include addr="#uname" or even
the wrong node.

If these operations detect the resource as running (after a cleanup)
then the digests won't match when the PE runs next, it will appear
the resource has changed definition and we will reload/restart it.