Test: pengine: new regression test for rebalancing unique clone instances
This is a variation of the bug-lf-2445 test where the clone has positive
(rather than 0) location preferences for each node.

As in bug-lf-2445, here a globally unique clone has more instances on one node
than optimal, the clone's stickiness is 0, and another node is available to run
the clone. The cluster should rebalance the clone instances across nodes.