Doc: pacemaker.service: Recommend not to limit tasks
Systemd v227 introduced the "pids" cgroup controller:
https://github.com/systemd/systemd/commit/03a7b521e

From systemd v228, it's enabled for all services by default, and the
global default of TasksMax was 512:
https://github.com/systemd/systemd/commit/9ded9cd14

The limit could be be easily hit since any processes/threads spawned
from the unit count, such as the "ocf" type of resources.

In systemd v235, the default TasksMax= has been effectively changed to
4915:
https://github.com/systemd/systemd/commit/79baeeb96

But it still could be a potential issue for cluster with heavy workload.

Generally it probably doesn't make much sense to set such a limit for a
resource manager like pacemaker.

This commit recommends the same as the docker project does:
https://github.com/moby/moby/commit/33a8ab29ed