Build: spec: make -libs-devel package dependencies arch-qualified
Beside the fact that *.so symlinks delivered with that package are
indeed satisfied with and only with arch-preserving -{,cluster-}libs
packages, it's also a matter of being disciplined per what distros
reasonably require:
https://fedoraproject.org/wiki/Packaging:Guidelines#Architecture-specific_Dependencies
(https://fedoraproject.org/wiki/Packaging:Guidelines#Requiring_Base_Package)

The syntactic provision at hand was introduced in rpm 4.6.0
(8+ years ago): http://rpm.org/user_doc/arch_dependencies.html

Also slightly reformat so as not to overflow reasonable line width
while grouping some related projects along.

Note that requirement on -cts hasn't been switched as that package
is practically noarch, except for lrmd_test, which should eventually
be fixed (mover over to -libs-devel?), so keep the state as-is for that
as a reminder.