Fix: attrd: ensure node name is broadcast at start-up (CLBZ#5330)
This fixes a regression introduced in 1.1.18.

Since c9d1c3cd, the crmd no longer explicitly clears the terminate and shutdown
node attributes at first join. An unwanted side effect of this was that the
attrd writer no longer reliably learned a joining node's name. If a node is
known only by its ID, the writer can not write its attributes to the CIB.

The worst outcome is that the joining node would be unable to shut down,
since the shutdown attribute would never trigger the policy engine. The window
was limited because the writer learns the node's name if a new attrd
election was required, or a node attribute was set locally on the joining node.

The fix is to set a new private attribute, #attrd-protocol, at attrd start-up,
with the supported attrd protocol version. This has the additional benefit of
allowing any node to determine the minimum supported protocol version across
all active cluster nodes.