Low: libpe_status: recover after failed demote when appropriate
Previously, the PE would always do a full stop after any failed demote.
However, if the resource needed recovery, the necessary start wouln't be
scheduled until the next transition.

Now, we skip the forced stop. It wasn't needed because native_create_actions()
does the right thing already. If the resource's next role is stopped, it will
schedule the right actions to get to stopped. If the resource's next role is
something else, it will force a stop and start for master/slave resources when
both a current node and a chosen node have been set (indicating recovery).