Test: PE: add tests for recovery after failed demote
There are two tests, one for recovering the resource in slave mode,
and one for recovering it in master mode.

The master test is a rare situation -- if the instance needed to be master, we
wouldn't have scheduled a demote. But this handles the situation when we get
the demote failure result and a new (winning) master score for the node at the
same time.