XML: technical debt: expose implemented particular-clone selectors
Likely due to messiness around schema versioning until it got settled
(e.g. validate-with="pacemaker-1.1" previously having the same role
as "pacemaker-next" nowadays) it happened that some internally supported
handling of selectors for particular clone instances in the constraints,
in particular:

1. /cib/configuration/constraints/rsc_colocation/@*[
       name() = 'rsc-instance' or name() = 'with-rsc-instance'
   ]

2. /cib/configuration/constraints/rsc_order[
       name() = 'first-instance' or name() = 'then-instance'
   ]

3. /cib/configuration/constraints/rsc_ticket/@rsc-instance

have never received proper finalization in terms of being permitted
in the proper versions of the schema.

Why this circumstance is deemed accidental rather than intended is
justified as follows:

A. support for 1. and 2. was added with 471036d1a and d1bfa1334,
   and for 3. with 6a743c386,
B. which all happened in the same timeframe, 2010-2011 (even before
   some comparative changes to be mentioned), and filling the same
   schema, pacemaker-1.1, as some other evolutions like:
   - per-node starting score: 570aacefa
   - fencing topology: 5b06f62e4
   - require-all attribute for resource_set: 8d39fe142,
C. that, unlike 1. - 3., got eventually promoted to pacemaker-1.2
   schema (respecting the order), just in time for 1.1.8 release:
   - b369f000a
   - 3c27492dc
   - 7c6f9446d,
D. hence it's natural to consider the same should have happened to
   1. - 3.

Therefore, retrospectively rectify the imposed formal boundaries by
allowing a little more -- said optional particular-clone selectors
where supported with the existing implementation already.  Also,
adapt the affected tests a bit.