Feature: stonithd,libcrmcommon: auto-calculate stonith-watchdog-timeout when requested
Before commit 8d08316, a non-positive stonith-watchdog-timeout meant that
Pacemaker would calculate an appropriate timeout based on SBD_WATCHDOG_TIMEOUT.

That did not allow users to disable watchdog-based fencing in Pacemaker when
sbd was in use, so 8d08316 changed nonpositive values to mean that.

However, there are use cases for each approach, so now, 0 (which is still the
default) means disable, while negative means auto-calculate.