XML: upgrade-2.10.xsl: re(drop) rsc_colocation/@score-attribute{,-mangle}
This is an incarnation of 30383cc5d commit that also provides the
necessary context, which won't be duplicated here.  It got subsequently
reverted in 74eae9e8f, as unfortunately it turned up to be possibly
causing schema upgrade issues.  So now, with the schema version making
a leap, accompanied with a dedicated upgrade conversion (as an XSLT
file), it's about the right time roll this out again, once and for
all now.  The act of making the change explicit boils down to mere
constraints-{next -> 3.0}.rng renaming (happend in 96d7ffedf), so the
only effective change is in the upgrade-2.10.xsl transformation recipe,
which is also accompanied with a unit test fitting the new test runner,
and update cibtr-2.rng with how cibtr:map grammar evolves.