Fix: tools: crm_master should always work on node attribute
Before ccbdb2a, crm_master would always set --node, thus ensuring crm_attribute
would treat the value as a node attribute. That commit removed that so that
crm_attribute could determine the local node name properly, but that introduced
an issue where the master value would be set as a cluster property instead of a
node attribute if --lifetime (or --node) was not set explicitly.

This fixes it by setting the default value of --lifetime explicitly.