XML: upgrade-*.xsl: bump the validate-with field from within stylesheet
This obsoletes doing the same on pacemaker side, but keep that code in
for a bit in case someone mixes the templates wildly (e.g. through
PCMK_schema_directory environment variable).

Logically, the upgrade always happens to the nearest higher major
version of the validating schema, no more, as arbitrary minor version
of the schema is required to be backward-compatible within the same
major one -- i.e. 1.x -> 2.0 and 2.x -> 3.0.

Also add a header with authorship and licensing info to upgrade-2.0.xsl
(only there), since:
- the file is slowly getting past the identity-transform-only phase
- substantially more sophistications is forthcoming, and they don't
  need to be spoiled with such a change as they are quite complex
  on their own