XML: upgrade-2.10.xsl: operation's "requires" [1/2]: drop/move around
See commit 014a543d53d.

Also make MapMsg helper, upgrade-detail.rng and cibtr-2.rng accommodate
the new move (@where) and value-based selector (@in-case-of) contexts
encoded in the upgrade "map".

In addition, introduce a possibility for parent/including grammar to,
optionally at its discretion, restrict nvpair/@name.  This comes handy
as we prohibit //operations/op/meta_attributes/nvpair[@name='requires'],
since it should no longer occur when the currently added parts of
upgrade are in effect; but we moreover generalize this overridding
mechanism with ability to define whole name+value pairs that can be
enumerated as a form of blacklist [known to come handy later on].