Refactor: tools: remove redundant code in crm_resource --list-cts
The code was made redundant by a regression in 1.1.5, via f398e48a which
incorrectly dropped an "else". However the code became incomplete anyway since
the introduction of "requires" in resource meta-attributes, and irrelevant
since "requires" was dropped from operation meta-attributs.

It hasn't been a problem since quorum requirement is synonymous with
non-stonith resource in the CTS generated configuration. Add a comment
describing a real fix, to be done later.