Build: configure: improve fatal warnings detection
Move the various bits of code that set $enable_fatal_warnings together,
for readability. The previous code had an ineffective system-specific test
that only worked because the it defaulted to yes later. Also, log a notice
if fatal warnings could not be enabled due to lack of (known) compiler support.