Build: configure: require GNU-compatible getopt when building
Some of Pacemaker's shell scripts use GNU getopt syntax for argument parsing,
and thus will fail when the first getopt on the PATH is something else (like
BSD getopt).

When configuring, search the entire PATH for a GNU-compatible getopt, exiting
with an error if none is found, and set a new substitution variable GETOPT_PATH
to its full path.