Build: spec: acknowledge Fedora getting rid of per-package ldconfig
It's natural considering the effectivity of doing so vs. once per whole
transaction (except for delicate cases where the scriptlet requires
the new version of the library or so).  This change should be
non-intrusive (probability of name clash of any sort is low, and
the original behaviour is preserved otherwise).

References:
https://fedoraproject.org/wiki/Changes/Removing_ldconfig_scriptlets