Low: controld: deprecate and ignore configdir parameter
DLM itself never supported an alternate location for configfs, and even the
controld agent had one hardcoded instance of the standard location.

Bump the resource agent version from 0.9 to 1.0, as it is widely used in
production.

The deprecation is intended only for backporting to the Pacemaker 1.1 release
series; the parameter will be removed entirely for the 2.0 release series.