XML: upgrade-2.10.xsl: space 2/4: extend HelperDenormalizedSpace
Now, the superfluous "marking" spaces are only to be emitted in the
"internal oracle" mode, imposing a new expectation that immediate
results gathered like that should not be passed to the output directly,
only to influence the decisions like whether the parent element should
be emitted at all so as not to remain empty.  This is currently not the
case, hence to preserve passing "./regression.sh -B" check for the time
being, set forcibly such a mode at relevant places that should really
be deriving the appropriate value from the callee's closure.