XML: upgrade-2.10.xsl: space 3/4: refactor InnerPass as a passable var.
Depending on the context, this can allow for optimizations (cf. passing
already computed "simulation" value from ProcessNonattrOpMetaAttributes
to the same template called without that flag) within
ProcessAttrOpMetaAttributes) or to discern cases when a template is
entered in an InnerSimulation mode directly from cases when it happens
privately (indicated with INNER-RECURSION value that is, in parallel,
getting introduced as a new convention here).