Low: libcrmcommon: check file/directory writability better
This replaces crm_is_writable() with a new similar function
pcmk__daemon_can_write() (starting a new convention of prefixing internal
library API functions with "pcmk__").

The old function was called with need_both=FALSE so that both user and group
write permissions were not required. However, it was implemented such that the
check would pass even if neither could write.

The new function fixes that and simplifies the interface by restricting it to
the only case we're interested in (either CRM_DAEMON_USER or CRM_DAEMON_GROUP
owns and can write the target). It also gives more detailed log messages when
something doesn't match.