Fix: libcrmcommon,pacemakerd: kernel task name is at most 15 characters
/proc/PID/status gives only up to 15 characters for the process name.
Change crm_procfs_process_info() to reflect this. Also, fix to handle
spaces in the name (though that's not an issue for current usage).

Similarly, only compare up to 15 characters against the result in pacemakerd,
which fixes a regression introduced in 2.0.0-rc3 making pacemakerd unable to
locate existing daemon processes at start-up.