Test: cts: install dummy systemd daemon via helper script
When bf0a2281 externalized the dummy daemon, it installed it via make.
This had the advantage of installing and cleaning once, rather than every test
run, and avoided leaving files behind if a test crashed. However, it broke
running tests from an uninstalled build directory.

The new approach is to have a helper script that installs the daemon,
have the tests call the helper script, and additionally offer a make target to
run the helper. This goes back to installing and cleaning every test run,
but the make target can be used to clean up after crashed tests.