Maint: make systemctl react to pacemaker-remote as with underscored one
This is a baby step in a long-term transition towards possibly dropping
the latter with some of the next major versions, presumably preceded
with the interim inversion of the main - alias roles (alternatively,
this symlink can be shipped right away to make it more robust, but
see https://github.com/systemd/systemd/issues/6338).
Note that the commit summary is not completely true, it only holds
once "pacemaker_remote" (in this and only this customary form) gets
enabled.

The idea is that the new spelling won't get actively promoted in the
short term, but later on, most common actions will work out-of-the-box
when dashed variant is used (either by user directly or by the
management layer); e.g. "is-enabled" will exit with 1 if the service
is not enabled regardless which variant is authoritative, only the
diagnostic messages may vary, and similarly with "status" when
coarse-grained decision is applied (zero vs. non-zero exit status).

That said, with either spelling, this is compatible with the
anticipated enable-then-start usage pattern under systemd supervision,
i.e. do not start without enabling first, and remains fully compatible
in whatever scenario when original underscored variant is used.