Build: spec: move old logs after upgrade
The main benefit is that the old logs will not otherwise get rotated (ever)
once an existing host is upgraded. Users who are upgrading would have to know
to remove the old logs manually if and when they no longer want them.

A second benefit is that if users notice the logs in /var/log first (or turn to
them out of habit), they might get frustrated when they don't see any new
messages come in.

The main potential downside is that upgrading, downgrading, then upgrading
again could result in loss of logs, but that seems a lesser problem than the
above.