XML: regression.sh: allow for granular test specifications
The granularity is test-prefix-based, hence can be narrowed down
to single test selection.  This prefix can now be optionally provided
right after the respective suite selector, with a slash as a delimiter.

For convenience (at least with "test2to3" currently), just the prefix
is needed, ".xml" extension is implicitly assumed, and likewise the
glob-style '*' wildcard is automatically appended at the right place
if missing.  It should be possible to specify multiple isolated
granular composite test selections even if they share the same
underlying suite (to be honest, this hasn't been tested).