Build: spec: add a simple retention of pre-existing log move targets
Arranged using coreutils-compatible "mv" implementation (we require
coreutils package for respective %post as opposed to a particular path,
as that would be against versatility [e.g. /bin vs. /usr/bin], but then
are relaxed about which "mv" is observed first per $PATH) instructed
to rename would-be-clobbered files with .rpmsave extension (per
convention & supporting discoverability), with a fallback to
unconditional clobbering of the pre-existing move targets (as it
used to be, and what will be used also when, e.g., a backup .rpmsave
file already exists and is set immutable).

Presence of the pre-2.0 log files clashing with 2.0 defaults is
unlikely, but better to be safe (consider also 2.0 release candidates
up to and including -rc3, or wild upgrade/downgrade scenarios).