Fix: libpe_status: find active instances properly according to requires
If a resource has "requires" set to "nothing" or "quorum", that means we can
properly start it elsewhere, even if the node believed to be initially running
the resource is unclean and waiting to be fenced.

Previously, if we did start the resource elsewhere before fencing completed,
the cluster would then consider the resource multiply active, and recover it.
Now, we don't consider such a resource multiply active if it's active on
only one clean node.

Status displays still show the resource as started on the unclean node, to give
the administrator a better idea of the actual situation. However, the clean
node will be considered the "current" node.