Refactor: libpe_status: new functions for finding resource's active nodes
Existing code often grabs rsc->running_on->data (i.e. the first node in the
list) as the resource's current node, and often uses
g_list_length(rsc->running_on).

However, if the resource is in the middle of a partial migration, the migration
source should be preferred as the current node. Also, if a resource has
"requires" set to "nothing" or "quorum", a clean, online node should be
preferred as the current node, and a caller should ignore unclean and offline
nodes when counting in certain cases.

These functions will allow those issues to be addressed in later commits.