Refactor: libpe_status: make clone unpack more efficient
Previously, when unpacking clone resource history, we would loop through all
possible clone instances until we found an instance active on the node, and
then if not found, we would loop through all possible instances again until we
found an inactive instance we could assign to the node.

Now, we only loop through the instances once, remembering the first inactive
instance. If we end up not finding an active instance, we return the remembered
inactive instance.