Low: scheduler: avoid unnecessary delay of start actions
If a node is being fenced, and we don't know the status of a resource there,
then any start of that resource must wait until the fencing is complete,
just in case it is active there.

However, the previous calculation would only check the known status of the
exact resource itself. If the resource is an anonymous clone instance, it is
sufficient if any instance of the clone is known on the node (with the
exception of partially active cloned groups).